+++ WooGym/CHANGELOG.md
@@ -1,15 +1,23 @@
-## [1.1.1] - 2023-01-06
+## [1.2.0] - 2024-01-15
+
+* Rename `AccountLandingPage` to `LoginPage`
+* Rename `AccountReigsterationPage` to `RegisterPage`
+* Update MinimumOSVersion to 12.0
+* Small refactor to project
+* Pubspec dependency updates
+
+## [1.1.1] - 2024-01-06
 
 * Update import paths
 * Add more translations
 * Pubspec dependency updates
 
-## [1.1.0] - 2023-01-05
+## [1.1.0] - 2024-01-05
 
 * Add more translations
 * Dart format
 * Pubspec dependency updates
 
-## [1.0.0] - 2023-01-04
+## [1.0.0] - 2024-01-04
 
 * Initial Release